llevmos el tema del menu de mensajes a otro archivo
